\name{HUG-package}
\alias{HUG-package}
\alias{HUG}
\docType{package}
\title{
\packageTitle{HUG}
}
\description{
\packageDescription{HUG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HUG}
\packageIndices{HUG}
}
\author{
\packageAuthor{HUG}

Maintainer: \packageMaintainer{HUG}
}
\references{
Liu, Jianyu, Haodong Wang, Wei Sun, Yufeng Liu. "Graphical Model Estimation for Single Cell RNA-seq Data." (2019+)
}
\keyword{graphical model, sample dependence, scRNA-seq}
